requirements.txt
discord.py
lyricsgenius
python-dotenv
spotipy
wavelink